fix issues with relative paths when bundling web apps
